requirements.txt
streamlit==1.45.1
fastapi==0.115.12
pydantic==2.11.6
uvicorn==0.34.3
transformers==4.41.2          # remains a solid, recent release
torch==2.3.0                  # latest stable GPU‑compatible release
pandas==2.2.2                 # recent stable
pymupdf==1.24.3               # current at project time
python-docx==1.1.2
fuzzywuzzy==0.18.0
sentence-transformers==2.6.1
gdown==5.1.0
python-dotenv==1.0.1
